pyproject.toml
[project]
name = "groups-evo-crewai"
version = "0.1.0"
description = "WhatsApp Group Manager and Summarizer"
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "crewai>=0.100.0",
    "crewai-tools>=0.32.1",
    "evolutionapi>=0.0.9",
    "streamlit>=1.41.1",
    "watchdog>=3.0.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.0",
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
    "plotly>=6.0.1",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "redis[hiredis]>=6.1.0",
    "hiredis>=3.2.1",
    "prometheus-client>=0.22.0",
    "psutil>=7.0.0",
    "structlog>=25.3.0",
    "click>=8.1.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--tb=short"
]

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "venv/*",
    "*/__pycache__/*",
    "*.pyc"
]

